test: mock the core read surface, enforce mocks for new RPC methods, and fix what that surfaced

+++ crates/edgefirst-client-py/src/lib.rs
@@ -1865,7 +1865,7 @@ impl Organization {
     }
 
     #[getter]
-    pub fn credits(&self) -> i64 {
+    pub fn credits(&self) -> f64 {
         self.0.credits()
     }
 }
